+++ src/main/java/duke/constant/EditType.java
@@ -1,3 +1,5 @@
+package duke.constant;
+
 public enum EditType {
     DONE,
     DELETE
