+++ src/main/java/duke/constant/TaskType.java
@@ -1,3 +1,5 @@
+package duke.constant;
+
 public enum TaskType {
     TODO,
     DEADLINE,
