+++ src/main/java/duke/parser/Parser.java
@@ -1,3 +1,18 @@
+package duke.parser;
+
+import duke.command.AddCommand;
+import duke.command.Command;
+import duke.command.EditCommand;
+import duke.command.ExitCommand;
+import duke.command.ListCommand;
+import duke.constant.EditType;
+import duke.constant.TaskType;
+import duke.exception.DukeException;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.ToDo;
+
 import java.util.List;
 import java.util.stream.Collectors;
